Update project.js
The previous code checked only the first row if the lines input is 1,but here all the rows are checked and if any row is allSame then the corresponding winning is updated. Similarly it takes care of situations where more than one row are allSame by giving the highest winning.